[장영진] 1주차 과제 3

+++ Section 1_준비/week1/과제 3 디버그하기/소스.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int getSequenceNumber() {
-    int a = 10;
+    double a = 10;
 
     for (int i = 0; i < 99; i++) {
         a /= 2;
